+++ CHANGELOG.md
@@ -6,6 +6,9 @@ CHANGELOG
 - Update pip/setuptools/wheel in virtual environment before installing dependencies
   [#5042](https://github.com/pulumi/pulumi/pull/5042)
 
+- Add ability to change a secrets provider for the current stack
+  [#5030](https://github.com/pulumi/pulumi/pull/5030)
+
 ## 2.7.1 (2020-07-22)
 
 - Fix logic to parse pulumi venv on github action
@@ -24,6 +27,7 @@ CHANGELOG
 
 - Add logic to parse pulumi venv on github action
   [#4994](https://github.com/pulumi/pulumi/pull/4994)
+ 
 
 - Better performance for stacks with many resources using the .NET SDK
   [#5015](https://github.com/pulumi/pulumi/pull/5015)